UPDATE: fix struct folder
