--- book/delivery/http/middleware/cors.go
@@ -1,10 +1,0 @@
-package middleware
-
-import "github.com/gin-gonic/gin"
-
-func Cors() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		c.Writer.Header().Add("Access-Control-Allow-Origin", "*")
-		c.Next()
-	}
-}
