+++ domain/book.go
@@ -11,20 +11,12 @@ type Book struct {
 	Author string `json:"author"`
 }
 
-type BookUsecase interface {
+type BookEntity interface {
 	Fetch(ctx context.Context) ([]Book, error)
 	GetByID(ctx context.Context, id string) (Book, error)
-	// Update(ctx context.Context, ar *Article) error
-	// GetByTitle(ctx context.Context, title string) (Article, error)
-	// Store(context.Context, *Article) error
-	// Delete(ctx context.Context, id int64) error
 }
 
 type BookRepository interface {
 	Fetch(ctx context.Context) (res []Book, err error)
 	GetByID(ctx context.Context, id string) (Book, error)
-	// GetByTitle(ctx context.Context, title string) (Article, error)
-	// Update(ctx context.Context, ar *Article) error
-	// Store(ctx context.Context, a *Article) error
-	// Delete(ctx context.Context, id int64) error
 }
